Move Equation class into a separate module, add documentation, remove EquationException, deprecate Equation.tangent_linear None return value
