+++ tlm_adjoint/equations.py
@@ -22,33 +22,25 @@ from .interface import check_space_types, check_space_types_conjugate_dual, \
     check_space_types_dual, conjugate_dual_space_type, function_assign, \
     function_axpy, function_comm, function_copy, function_dtype, \
     function_get_values, function_global_size, function_id, function_inner, \
-    function_is_alias, function_is_checkpointed, function_local_indices, \
-    function_local_size, function_new, function_new_conjugate_dual, \
-    function_replacement, function_set_values, function_space, \
-    function_space_type, function_sum, function_update_caches, \
-    function_update_state, function_zero, is_function, \
+    function_local_indices, function_local_size, function_new, \
+    function_new_conjugate_dual, function_replacement, function_set_values, \
+    function_space, function_space_type, function_sum, \
+    function_update_caches, function_zero, is_function, \
     no_space_type_checking, space_new
 
 from .adjoint import AdjointModelRHS
-from .alias import WeakAlias, gc_disabled
-from .manager import manager as _manager
-from .manager import paused_manager, restore_manager, set_manager
+from .alias import WeakAlias
+from .equation import Equation, Referrer
 
 from collections.abc import Sequence
 import copy
 import inspect
 import logging
 import numpy as np
-from operator import itemgetter
 import warnings
-import weakref
 
 __all__ = \
     [
-        "EquationException",
-
-        "Equation",
-
         "ControlsMarker",
         "FunctionalMarker",
 
@@ -88,511 +80,6 @@ __all__ = \
         "SumRHS",
         "SumSolver"
     ]
-
-
-class EquationException(Exception):  # noqa: N818
-    def __init__(self, *args, **kwargs):
-        warnings.warn("EquationException is deprecated",
-                      DeprecationWarning, stacklevel=2)
-        super().__init__(*args, **kwargs)
-
-
-class Referrer:
-    _id_counter = [0]
-
-    def __init__(self, referrers=None):
-        if referrers is None:
-            referrers = []
-
-        self._id = self._id_counter[0]
-        self._id_counter[0] += 1
-        self._referrers = weakref.WeakValueDictionary()
-        self._references_dropped = False
-
-        self.add_referrer(*referrers)
-
-    def id(self):
-        return self._id
-
-    @gc_disabled
-    def add_referrer(self, *referrers):
-        if self._references_dropped:
-            raise RuntimeError("Cannot call add_referrer method after "
-                               "_drop_references method has been called")
-        for referrer in referrers:
-            referrer_id = referrer.id()
-            assert self._referrers.get(referrer_id, referrer) is referrer
-            self._referrers[referrer_id] = referrer
-
-    @gc_disabled
-    def referrers(self):
-        referrers = {}
-        remaining_referrers = {self.id(): self}
-        while len(remaining_referrers) > 0:
-            referrer_id, referrer = remaining_referrers.popitem()
-            if referrer_id not in referrers:
-                referrers[referrer_id] = referrer
-                for child in tuple(referrer._referrers.valuerefs()):
-                    child = child()
-                    if child is not None:
-                        child_id = child.id()
-                        if child_id not in referrers and child_id not in remaining_referrers:  # noqa: E501
-                            remaining_referrers[child_id] = child
-        return tuple(e[1] for e in sorted(tuple(referrers.items()),
-                                          key=itemgetter(0)))
-
-    def _drop_references(self):
-        if not self._references_dropped:
-            self.drop_references()
-            self._references_dropped = True
-
-    def drop_references(self):
-        raise NotImplementedError("Method not overridden")
-
-
-class Equation(Referrer):
-    def __init__(self, X, deps, nl_deps=None,
-                 *, ic_deps=None, ic=None,
-                 adj_ic_deps=None, adj_ic=None,
-                 adj_type="conjugate_dual"):
-        """
-        An equation. The equation is expressed in the form:
-            F ( X, y_0, y_1, ... ) = 0,
-        where X is the equation solution and F is a residual function.
-        Information regarding F is provided by methods which should be
-        overridden as required by derived classes.
-
-        Arguments:
-
-        X        A function, or a sequence of functions. The solution to the
-                 equation.
-        deps     A sequence of dependencies, which must include the solution
-                 itself.
-        nl_deps  (Optional) A sequence of non-linear dependencies. Must be a
-                 subset of deps. Defaults to deps.
-        ic_deps  (Optional) A sequence of dependencies whose initial value
-                 should be available prior to solving the forward equation.
-                 Must be a subset of X.
-        ic       (Optional) If true then ic_deps is set equal to X. Defaults to
-                 true if ic_deps is None, and false otherwise.
-        adj_ic_deps  (Optional) A sequence of dependencies whose adjoint value
-                     should be available prior to solving the adjoint equation.
-                     Must be a subset of X.
-        adj_ic       (Optional) If true then adj_ic_deps is set equal to X.
-                     Defaults to true if adj_ic_deps is None, and false
-                     otherwise.
-        adj_type  (Optional) "primal" or "conjugate_dual", or a sequence of
-                  these, defining whether elements of the adjoint are in the
-                  conjugate dual space associated with corresponding elements
-                  of X.
-        """
-
-        if is_function(X):
-            X = (X,)
-        X_ids = set(map(function_id, X))
-        dep_ids = {function_id(dep): i for i, dep in enumerate(deps)}
-        for x in X:
-            if not is_function(x):
-                raise ValueError("Solution must be a function")
-            if not function_is_checkpointed(x):
-                raise ValueError("Solution must be checkpointed")
-            if function_is_alias(x):
-                raise ValueError("Solution cannot be an alias")
-            if function_id(x) not in dep_ids:
-                raise ValueError("Solution must be a dependency")
-
-        if len(dep_ids) != len(deps):
-            raise ValueError("Duplicate dependency")
-        for dep in deps:
-            if function_is_alias(dep):
-                raise ValueError("Dependency cannot be an alias")
-
-        if nl_deps is None:
-            nl_deps = tuple(deps)
-        nl_dep_ids = set(map(function_id, nl_deps))
-        if len(nl_dep_ids) != len(nl_deps):
-            raise ValueError("Duplicate non-linear dependency")
-        for dep in nl_deps:
-            if function_id(dep) not in dep_ids:
-                raise ValueError("Non-linear dependency is not a dependency")
-
-        if ic_deps is None:
-            ic_deps = []
-            if ic is None:
-                ic = True
-        else:
-            if ic is None:
-                ic = False
-        ic_dep_ids = set(map(function_id, ic_deps))
-        if len(ic_dep_ids) != len(ic_deps):
-            raise ValueError("Duplicate initial condition dependency")
-        for dep in ic_deps:
-            if function_id(dep) not in X_ids:
-                raise ValueError("Initial condition dependency is not a "
-                                 "solution")
-        if ic:
-            ic_deps = list(X)
-
-        if adj_ic_deps is None:
-            adj_ic_deps = []
-            if adj_ic is None:
-                adj_ic = True
-        else:
-            if adj_ic is None:
-                adj_ic = False
-        adj_ic_dep_ids = set(map(function_id, adj_ic_deps))
-        if len(adj_ic_dep_ids) != len(adj_ic_deps):
-            raise ValueError("Duplicate adjoint initial condition dependency")
-        for dep in adj_ic_deps:
-            if function_id(dep) not in X_ids:
-                raise ValueError("Adjoint initial condition dependency is not "
-                                 "a solution")
-        if adj_ic:
-            adj_ic_deps = list(X)
-
-        if adj_type in ["primal", "conjugate_dual"]:
-            adj_type = tuple(adj_type for x in X)
-        elif isinstance(adj_type, Sequence):
-            if len(adj_type) != len(X):
-                raise ValueError("Invalid adjoint type")
-        else:
-            raise ValueError("Invalid adjoint type")
-        for adj_x_type in adj_type:
-            if adj_x_type not in ["primal", "conjugate_dual"]:
-                raise ValueError("Invalid adjoint type")
-
-        super().__init__()
-        self._X = tuple(X)
-        self._deps = tuple(deps)
-        self._nl_deps = tuple(nl_deps)
-        self._ic_deps = tuple(ic_deps)
-        self._adj_ic_deps = tuple(adj_ic_deps)
-        self._adj_X_type = tuple(adj_type)
-
-    _reset_adjoint_warning = True
-    _initialize_adjoint_warning = True
-    _finalize_adjoint_warning = True
-
-    def __init_subclass__(cls, **kwargs):
-        super().__init_subclass__(**kwargs)
-
-        if hasattr(cls, "reset_adjoint"):
-            if cls._reset_adjoint_warning:
-                warnings.warn("Equation.reset_adjoint method is deprecated",
-                              DeprecationWarning, stacklevel=2)
-        else:
-            cls._reset_adjoint_warning = False
-            cls.reset_adjoint = lambda self: None
-
-        if hasattr(cls, "initialize_adjoint"):
-            if cls._initialize_adjoint_warning:
-                warnings.warn("Equation.initialize_adjoint method is "
-                              "deprecated",
-                              DeprecationWarning, stacklevel=2)
-        else:
-            cls._initialize_adjoint_warning = False
-            cls.initialize_adjoint = lambda self, J, nl_deps: None
-
-        if hasattr(cls, "finalize_adjoint"):
-            if cls._finalize_adjoint_warning:
-                warnings.warn("Equation.finalize_adjoint method is deprecated",
-                              DeprecationWarning, stacklevel=2)
-        else:
-            cls._finalize_adjoint_warning = False
-            cls.finalize_adjoint = lambda self, J: None
-
-        adj_solve_sig = inspect.signature(cls.adjoint_jacobian_solve)
-        if tuple(adj_solve_sig.parameters.keys()) in [("self", "nl_deps", "b"),
-                                                      ("self", "nl_deps", "B")]:  # noqa: E501
-            warnings.warn("Equation.adjoint_jacobian_solve(self, nl_deps, b/B) "  # noqa: E501
-                          "method signature is deprecated",
-                          DeprecationWarning, stacklevel=2)
-
-            def adjoint_jacobian_solve(self, adj_X, nl_deps, B):
-                return adjoint_jacobian_solve_orig(self, nl_deps, B)
-            adjoint_jacobian_solve_orig = cls.adjoint_jacobian_solve
-            cls.adjoint_jacobian_solve = adjoint_jacobian_solve
-
-    def drop_references(self):
-        self._X = tuple(function_replacement(x) for x in self._X)
-        self._deps = tuple(function_replacement(dep) for dep in self._deps)
-        self._nl_deps = tuple(function_replacement(dep)
-                              for dep in self._nl_deps)
-        self._ic_deps = tuple(function_replacement(dep)
-                              for dep in self._ic_deps)
-        self._adj_ic_deps = tuple(function_replacement(dep)
-                                  for dep in self._adj_ic_deps)
-
-    def x(self):
-        x, = self._X
-        return x
-
-    def X(self, m=None):
-        if m is None:
-            return self._X
-        else:
-            return self._X[m]
-
-    def dependencies(self):
-        return self._deps
-
-    def nonlinear_dependencies(self):
-        return self._nl_deps
-
-    def initial_condition_dependencies(self):
-        return self._ic_deps
-
-    def adjoint_initial_condition_dependencies(self):
-        return self._adj_ic_deps
-
-    def adj_x_type(self):
-        adj_x_type, = self.adj_X_type()
-        return adj_x_type
-
-    def adj_X_type(self, m=None):
-        if m is None:
-            return self._adj_X_type
-        else:
-            return self._adj_X_type[m]
-
-    def new_adj_x(self):
-        adj_x, = self.new_adj_X()
-        return adj_x
-
-    def new_adj_X(self, m=None):
-        if m is None:
-            return tuple(self.new_adj_X(m) for m in range(len(self.X())))
-        else:
-            return function_new(self.X(m), rel_space_type=self.adj_X_type(m))
-
-    def _pre_process(self, manager=None, annotate=None):
-        if manager is None:
-            manager = _manager()
-        for dep in self.initial_condition_dependencies():
-            manager.add_initial_condition(dep, annotate=annotate)
-
-    def _post_process(self, manager=None, annotate=None, tlm=None):
-        if manager is None:
-            manager = _manager()
-        manager.add_equation(self, annotate=annotate, tlm=tlm)
-
-    @restore_manager
-    def solve(self, manager=None, annotate=None, tlm=None):
-        """
-        Solve the equation.
-
-        Arguments:
-
-        manager   (Optional) The equation manager.
-        annotate  (Optional) Whether the equation should be recorded.
-        tlm       (Optional) Whether to derive (and solve) associated
-                  tangent-linear equations.
-        """
-
-        if manager is not None:
-            set_manager(manager)
-
-        self._pre_process(annotate=annotate)
-
-        with paused_manager():
-            self.forward(self.X())
-
-        self._post_process(annotate=annotate, tlm=tlm)
-
-    def forward(self, X, deps=None):
-        """
-        Solve the equation. The manager is stopped when this method is called.
-
-        Arguments:
-
-        X     A sequence of functions. The solution, which should be set by
-              this method.
-        deps  (Optional) A sequence of functions defining the values of
-              dependencies.
-        """
-
-        function_update_caches(*self.dependencies(), value=deps)
-        self.forward_solve(X[0] if len(X) == 1 else X, deps=deps)
-        function_update_state(*X)
-        function_update_caches(*self.X(), value=X)
-
-    def forward_solve(self, X, deps=None):
-        """
-        Solve the equation. The manager is stopped when this method is called.
-
-        The form:
-            forward_solve(self, x, deps=None)
-        should be used for equations which solve for a single function.
-
-        Arguments:
-
-        x/X   The solution, which should be set by this method.
-        deps  (Optional) A sequence of functions defining the values of
-              dependencies. self.dependencies() should be used if this is not
-              supplied.
-        """
-
-        raise NotImplementedError("Method not overridden")
-
-    def adjoint(self, J, adj_X, nl_deps, B, dep_Bs):
-        """
-        Solve the adjoint equation with the given right-hand-side, and subtract
-        adjoint terms from other adjoint right-hand-sides.
-
-        Arguments:
-
-        J          Adjoint model functional.
-        adj_X      A sequence of functions defining the initial guess for the
-                   adjoint solve, or None if the Equation does not accept an
-                   initial guess. May be modified or returned by this method.
-        nl_deps    A sequence of functions defining the values of non-linear
-                   dependencies.
-        B          A sequence of functions defining the right-hand-side, which
-                   may be modified or returned by this method.
-        dep_Bs     Dictionary of dep_index: dep_B pairs, where each dep_B is an
-                   AdjointRHS which should be updated by subtracting derivative
-                   information computed by differentiating with respect to
-                   self.dependencies()[dep_index].
-
-        Returns the solution of the adjoint equation as a tuple of functions.
-        The result will not be modified by calling code.
-        """
-
-        function_update_caches(*self.nonlinear_dependencies(), value=nl_deps)
-        self.initialize_adjoint(J, nl_deps)
-
-        if adj_X is not None and len(adj_X) == 1:
-            adj_X = adj_X[0]
-        adj_X = self.adjoint_jacobian_solve(
-            adj_X, nl_deps, B[0] if len(B) == 1 else B)
-        if adj_X is not None:
-            self.subtract_adjoint_derivative_actions(adj_X, nl_deps, dep_Bs)
-
-            if is_function(adj_X):
-                adj_X = (adj_X,)
-
-            for m, adj_x in enumerate(adj_X):
-                check_space_types(adj_x, self.X(m),
-                                  rel_space_type=self.adj_X_type(m))
-
-        self.finalize_adjoint(J)
-
-        if adj_X is None:
-            return None
-        else:
-            return tuple(adj_X)
-
-    def adjoint_cached(self, J, adj_X, nl_deps, dep_Bs):
-        """
-        Subtract adjoint terms from other adjoint right-hand-sides.
-
-        Arguments:
-
-        J          Adjoint model functional.
-        adj_X      A sequence of functions defining the adjoint solution.
-        nl_deps    A sequence of functions defining the values of non-linear
-                   dependencies.
-        dep_Bs     Dictionary of dep_index: dep_B pairs, where each dep_B is an
-                   AdjointRHS which should be updated by subtracting derivative
-                   information computed by differentiating with respect to
-                   self.dependencies()[dep_index].
-        """
-
-        function_update_caches(*self.nonlinear_dependencies(), value=nl_deps)
-        self.initialize_adjoint(J, nl_deps)
-
-        if len(adj_X) == 1:
-            adj_X = adj_X[0]
-        self.subtract_adjoint_derivative_actions(adj_X, nl_deps, dep_Bs)
-
-        self.finalize_adjoint(J)
-
-    def adjoint_derivative_action(self, nl_deps, dep_index, adj_X):
-        """
-        Return the action of the adjoint of a derivative of the RHS.
-
-        The return value will not be modified by calling code.
-
-        The form:
-            adjoint_derivative_action(self, nl_deps, dep_index, adj_x)
-        should be used for equations which solve for a single function.
-
-        Arguments:
-
-        nl_deps      A sequence of functions defining the values of non-linear
-                     dependencies.
-        dep_index    The index of the dependency in self.dependencies() with
-                     respect to which a derivative should be taken.
-        adj_x/adj_X  A function or sequence of functions on which the adjoint
-                     of the derivative acts.
-        """
-
-        raise NotImplementedError("Method not overridden")
-
-    def subtract_adjoint_derivative_actions(self, adj_X, nl_deps, dep_Bs):
-        """
-        Subtract adjoint derivative actions from adjoint right-hand-sides.
-        Lower level than adjoint_derivative_action, but can be overridden for
-        optimization, and can be defined in place of defining an
-        adjoint_derivative_action method.
-
-        The form:
-            subtract_adjoint_derivative_actions(self, adj_x, nl_deps, dep_Bs)
-        should be used for equations which solve for a single function.
-
-        Arguments:
-
-        adj_x/adj_X  A function or sequence of functions on which the adjoint
-                     of the derivatives act.
-        nl_deps      A sequence of functions defining the values of non-linear
-                     dependencies.
-        dep_Bs       Dictionary of dep_index: dep_B pairs, where each dep_B is
-                     an AdjointRHS which should be updated by subtracting
-                     derivative information computed by differentiating with
-                     respect to self.dependencies()[dep_index].
-        """
-
-        for dep_index, dep_B in dep_Bs.items():
-            dep_B.sub(self.adjoint_derivative_action(nl_deps, dep_index,
-                                                     adj_X))
-
-    def adjoint_jacobian_solve(self, adj_X, nl_deps, B):
-        """
-        Solve an adjoint equation, returning the result. The result will not be
-        modified by calling code. A return value of None can be used to
-        indicate that the solution is zero.
-
-        The form:
-            adjoint_jacobian_solve(self, adj_x, nl_deps, b)
-        should be used for equations which solve for a single function.
-
-        Arguments:
-
-        adj_x/adj_X    Initial guess for the adjoint solve, or None if the
-                       Equation does not accept an initial guess. May be
-                       modified or returned by this method.
-        nl_deps        A sequence of functions defining the values of
-                       non-linear dependencies.
-        b/B            The right-hand-side. May be modified or returned by this
-                       method.
-        """
-
-        raise NotImplementedError("Method not overridden")
-
-    def tangent_linear(self, M, dM, tlm_map):
-        """
-        Return an Equation corresponding to a tangent linear equation,
-        computing derivatives with respect to the control M with direction
-        dM.
-
-        Arguments:
-
-        M        A sequence of functions defining the control.
-        dM       A sequence of functions defining the direction.
-        tlm_map  The TangentLinearMap.
-        """
-
-        raise NotImplementedError("Method not overridden")
 
 
 class ControlsMarker(Equation):
@@ -1306,6 +793,9 @@ class FixedPointSolver(Equation, CustomNormSq):
         for eq in self._eqs:
             tlm_eq = eq.tangent_linear(M, dM, tlm_map)
             if tlm_eq is None:
+                warnings.warn("Equation.tangent_linear should return an "
+                              "Equation",
+                              DeprecationWarning)
                 tlm_eq = ZeroAssignment([tlm_map[x] for x in eq.X()])
             tlm_eqs.append(tlm_eq)
         return FixedPointSolver(
